Got all squares to work, now only need detection

+++ trianglesurface.h
@@ -19,8 +19,8 @@ public:
     void construct_plane();
     std::vector<gsml::Vertex>& get_vertices() { return mVertices; } // 191120
     float xmin=0, xmax=1.0f, ymin=0.0f, ymax=1.0f, h=0.25f;
-    float xMinimum = 0, xMaximum = 1, yMinimum = 0, yMaximum=1, zMinimum = 0, zMaximum=1;
-    float numberofPointsInside;
+    double xMinimum = 0, xMaximum = 1, yMinimum = 0, yMaximum=1, zMinimum = 0, zMaximum=1;
+    int numberofPointsInside;
     gsml::Vertex vertex2;
     std::vector<gsml::Vertex> Vertices;
 };
